Coalesce some similar integration tests, add failing tests for whitespace
