+++ tests/integration/html-test.js
@@ -1,7 +1,34 @@
+/*global QUnit*/
 import Emblem from '../../emblem';
 import { w } from '../support/utils';
 
-QUnit.module("html multiple lines");
+QUnit.module("html: single line");
+
+QUnit.test("element only", function(assert){
+  assert.compilesTo("p", "<p></p>");
+});
+
+QUnit.test("with text", function(assert){
+  assert.compilesTo("p Hello", "<p>Hello</p>");
+});
+
+QUnit.test("with more complex text", function(assert){
+  assert.compilesTo(
+    "p Hello, how's it going with you today?",
+    "<p>Hello, how's it going with you today?</p>"
+  );
+});
+
+QUnit.test("with trailing space", function(assert){
+  assert.compilesTo("p Hello   ", "<p>Hello   </p>");
+});
+
+QUnit.test("can start with angle bracket html", function(assert){
+  var emblem = "<span>Hello</span>";
+  assert.compilesTo(emblem, "<span>Hello</span>");
+});
+
+QUnit.module("html: multiple lines");
 
 QUnit.test("two lines", function(assert){
   var emblem = w(
@@ -31,8 +58,8 @@ QUnit.test("three lines w/ embedded html", function(assert){
 
 QUnit.test("indentation doesn't need to match starting inline content's", function(assert){
   var emblem = w(
-    "span Hello,",
-    "  How are you?"
+    "  span Hello,",
+    "    How are you?"
   );
   assert.compilesTo(emblem, "<span>Hello, How are you?</span>");
 });
@@ -84,7 +111,7 @@ QUnit.test("w/ mustaches", function(assert){
     "<div><span>Hello, {{foo}} are you? Excellent.</span></div>");
 });
 
-// This test seems to test nonsense syntax
+// FIXME: This test seems to test nonsense syntax ?
 QUnit.test("w/ block mustaches", function(assert){
   var emblem = w(
     "p Hello, #{ sally | Hello},",
@@ -114,4 +141,39 @@ QUnit.test("can start with angle bracket html and go to multiple lines", functio
     "      what's up?</span>"
   );
   assert.compilesTo(emblem, "<span>Hello dude,\nwhat's up?</span>");
+});
+
+QUnit.module("html: nested");
+
+test("basic", function(assert){
+  var emblem = w(
+    "p",
+    "  span Hello",
+    "  strong Hi",
+    "div",
+    "  p Hooray"
+  );
+  assert.compilesTo(emblem,
+    '<p><span>Hello</span><strong>Hi</strong></p><div><p>Hooray</p></div>');
+});
+
+test("empty nest", function(assert){
+  var emblem = w(
+    "p",
+    "  span",
+    "    strong",
+    "      i"
+  );
+  assert.compilesTo(emblem, '<p><span><strong><i></i></strong></span></p>');
+});
+
+test("empty nest w/ attribute shorthand", function(assert){
+  var emblem = w(
+    "p.woo",
+    "  span#yes",
+    "    strong.no.yes",
+    "      i"
+  );
+  assert.compilesTo(emblem,
+    '<p class="woo"><span id="yes"><strong class="no yes"><i></i></strong></span></p>');
 });
