+++ tests/integration/mustaches-test.js
@@ -123,6 +123,38 @@ test('hash stache in text line', function(assert){
   assert.compilesTo(emblem, 'bork {{bar}}');
 });
 
+QUnit.module("mustache: hash brace syntax, #{}");
+
+test('acts like {{}}', function(assert){
+  var emblem = "span Yo #{foo}, I herd.";
+  assert.compilesTo(emblem,
+    "<span>Yo {{foo}}, I herd.</span>");
+});
+
+test('can start inline content', function(assert){
+  var emblem = "span #{foo}, I herd.";
+  assert.compilesTo(emblem,
+    "<span>{{foo}}, I herd.</span>");
+});
+
+test('can end inline content', function(assert){
+  var emblem = "span I herd #{foo}";
+  assert.compilesTo(emblem,
+    "<span>I herd {{foo}}</span>");
+});
+
+test("doesn't screw up parsing when # used in text nodes", function(assert){
+  var emblem = "span OMG #YOLO";
+  assert.compilesTo(emblem,
+    "<span>OMG #YOLO</span>");
+});
+
+test("# can be only thing on line", function(assert){
+  var emblem = "span #";
+  assert.compilesTo(emblem,
+    "<span>#</span>");
+});
+
 /*
 QUnit.module("mustache helpers");
